# src/main/java/de/gematik/zeta/testfachdienst/model/HelloZetaResource.java
/*-
 * #%L
 * ZETA Testfachdienst
 * %%
 * (C) achelos GmbH, 2025, licensed for gematik GmbH
 * %%
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * *******
 *
 * For additional notes and disclaimer from gematik and in case of changes by gematik
 * find details in the "Readme" file.
 * #L%
 */

package de.gematik.zeta.testfachdienst.model;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Getter;
import lombok.Setter;

/**
 * DTO that holds a greeting message returned by the ZETA showcase endpoint.
 */
@Schema(
    name = "HelloZetaResource",
    description = "Simple greeting payload returned by Hello ZETA endpoints")
public class HelloZetaResource {

  @Getter
  @Setter
  @Schema(
      description = "Greeting text returned to the caller",
      example = "Hello ZETA!")
  private String message;

  /**
   * Create an empty greeting resource instance for frameworks that require a default constructor.
   */
  public HelloZetaResource() {
  }

  /**
   * Create a resource initialized with the supplied greeting message.
   *
   * @param message greeting that should be sent to the caller
   */
  public HelloZetaResource(String message) {
    this.message = message;
  }
}